+++ apps/jotai-form-example/src/features/unit-events-page/UnitEventsTable.tsx
@@ -34,7 +34,7 @@ export function UnitEventsTable() {
     setNewRows([])
   }
 
-  console.log(newRows)
+  console.log(data)
 
   return (
     <div>
