+++ apps/jotai-form-example/src/features/unit-events-page/components/ActionsCell.tsx
@@ -4,7 +4,7 @@ import {
   useGetIsNew,
   useHasRemoteChange,
   useSetFormMeta,
-} from '@equinor/apollo-utils'
+} from '@equinor/apollo-components'
 import { Button, Icon, Progress } from '@equinor/eds-core-react'
 import { close, edit, save } from '@equinor/eds-icons'
 import { CellContext } from '@tanstack/react-table'
@@ -12,12 +12,9 @@ import { useFormContext, useFormState } from 'react-hook-form'
 import { useUnitEventMutation } from '../hooks/useUnitEventMutation'
 import { UnitEvent } from '../types'
 import { useUnitEventFormContext } from '../UnitEventsFormContext'
-import { unitEventFormUtils } from '../utils'
 
 export function ActionsCell({ row }: CellContext<UnitEvent, unknown>) {
-  const pokemonFormState = unitEventFormUtils.useFormState({ id: row.id })
-  const { initializeForm, resetForm } = unitEventFormUtils.useFormMutations({ id: row.id })
-  const mutatePokemon = useUnitEventMutation(row.original)
+  const mutatePokemon = useUnitEventMutation()
 
   const { setEditMode } = useUnitEventFormContext()
   const { handleSubmit } = useFormContext<UnitEvent>()
@@ -43,6 +40,7 @@ export function ActionsCell({ row }: CellContext<UnitEvent, unknown>) {
 
         const isNew = getIsNew()
         console.log(`Mutating ${isNew ? 'new' : 'updated'} Unit Event`, newValues)
+        mutatePokemon(newValues)
         // if (isNew) await mutatePost([{ valuesBeforeChange, valuesAfterChange }])
         // else await mutatePut([{ valuesBeforeChange, valuesAfterChange }])
