+++ apps/jotai-form-example/src/features/unit-events-page/hooks/useUnitEventMutation.ts
@@ -1,20 +1,18 @@
-import { FormState } from '@equinor/apollo-utils'
 import { useQueryClient } from '@tanstack/react-query'
 import { UnitEvent } from 'mock-data'
-import { unitEventFormUtils } from '../utils'
 
-export function useUnitEventMutation(unitEvent: UnitEvent) {
+export function useUnitEventMutation() {
   const queryClient = useQueryClient()
-  const { resetForm } = unitEventFormUtils.useFormMutations(unitEvent)
 
-  return (formState: FormState<UnitEvent>) => {
+  return (newValue: UnitEvent) => {
     queryClient.setQueryData(['allUnitEvents'], (currentUnitEvent?: UnitEvent[]) => {
+      console.log(currentUnitEvent)
+
       if (!currentUnitEvent) return
       return currentUnitEvent.map((unitEvent) => {
-        if (unitEvent.id === formState.values.id) return formState.values
+        if (unitEvent.id === newValue.id) return newValue
         return unitEvent
       })
     })
-    resetForm()
   }
 }
