+++ packages/apollo-components/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@equinor/apollo-components",
-  "version": "3.1.2",
+  "version": "3.1.2-beta.0",
   "license": "MIT",
   "main": "./dist/index.js",
   "module": "./dist/index.mjs",
@@ -29,7 +29,8 @@
     "@equinor/eds-core-react": "^0.27.0",
     "styled-components": "^5.3.7",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.43.8"
   },
   "devDependencies": {
     "@storybook/addon-essentials": "^7.0.0-alpha.38",
