+++ packages/apollo-components/src/form-cells/EditableDropdownCell.tsx
@@ -2,8 +2,8 @@ import { Autocomplete } from '@equinor/eds-core-react'
 import { CellContext } from '@tanstack/react-table'
 import { Controller, useFormContext } from 'react-hook-form'
 import styled from 'styled-components'
+import { TypographyCustom } from '../cells'
 import { FormMeta, useEditMode } from '../form-meta'
-import { TypographyCustom } from './TypographyCustom'
 
 export interface Option {
   label: string
