+++ packages/apollo-components/src/form-cells/EditableTextFieldCell.tsx
@@ -2,8 +2,8 @@ import { TextField } from '@equinor/eds-core-react'
 import { CellContext } from '@tanstack/react-table'
 import { Controller, FieldPath, useFormContext } from 'react-hook-form'
 import styled from 'styled-components'
+import { TypographyCustom } from '../cells'
 import { FormMeta, useEditMode } from '../form-meta'
-import { TypographyCustom } from './TypographyCustom'
 import { getHelperTextProps } from './utils'
 
 export function EditableTextFieldCell<T extends FormMeta>(context: CellContext<T, unknown>) {
