+++ packages/apollo-components/src/index.ts
@@ -1,3 +1,5 @@
 export { AppShell, AppSidebar } from './AppShell'
 export * from './cells'
 export * from './DataTable'
+export * from './form-cells'
+export * from './form-meta'
