+++ packages/apollo-utils/src/index.ts
@@ -1,4 +1,2 @@
-export * from './edit-cells'
-export * from './form-meta'
 export * from './jotai-form'
 export * from './zod-validation'
